+++ TechnicalTask/src/main/java/com/malykhnik/technicaltask/service/TransferMoneyService.java
@@ -1,5 +1,7 @@
 package com.malykhnik.technicaltask.service;
 
+import java.math.BigDecimal;
+
 public interface TransferMoneyService {
-    void transfer(Long userFromId, Long UserToId, Double money);
+    void transfer(Long userFromId, Long UserToId, BigDecimal money);
 }
